[HER-17]Follow-up Rename directory
