+++ src/detection/card/action/container.jsx
@@ -2,9 +2,9 @@ import React from 'react'
 import { CardActions as MaterialCardActions, Grid } from '@material-ui/core'
 import PropTypes from 'prop-types'
 
-import { useStyles } from 'detection/detection-card/action/style'
+import { useStyles } from 'detection/card/action/style'
 
-export const CardAction = ({ children }) => <MaterialCardActions className={useStyles().root}>
+export const CardAction = ({ children }) => <MaterialCardActions classes={useStyles()}>
   <Grid container justify={'space-around'} alignContent={'center'}>
     {children}
   </Grid>
