+++ src/detection/card/action/style.js
