+++ src/detection/card/content/container/box.jsx
@@ -2,9 +2,9 @@ import React from 'react'
 import { Box as MaterialBox } from '@material-ui/core'
 import PropTypes from 'prop-types'
 
-import { useStyles } from 'detection/detection-card/content/container/box-style'
+import { useStyles } from 'detection/card/content/container/box-style'
 
-export const Box = ({ children }) => <MaterialBox className={useStyles().root}>
+export const Box = ({ children }) => <MaterialBox classes={useStyles()}>
   {children}
 </MaterialBox>
