+++ src/detection/card/content/container/grid-style.js
