+++ src/detection/card/content/container/style.js
