+++ src/detection/card/content/items/items-style.js
