+++ src/detection/card/desktop-style.js
