+++ src/detection/card/mobile-style.js
