+++ src/detection/card/title/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { CardHeader } from '@material-ui/core'
+
+import { useStyles } from 'detection/card/title/style'
+
+export const Title = () => <CardHeader title={'Detection name'} classes={useStyles()} />
