--- src/detection/detection-card/title/index.jsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import { CardHeader } from '@material-ui/core'
-
-import { useStyles } from 'detection/detection-card/title/style'
-
-export const Title = () => <CardHeader title={'Detection name'} className={useStyles().root} />
